+++ database/migrations/2018_12_21_034417_add_address_and_shipping_to_orders_table.php
@@ -14,8 +14,8 @@ class AddAddressAndShippingToOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->unsignedInteger('address_id')->after('user_id')->index();
-            $table->unsignedInteger('shipping_method_id')->after('address_id')->index();
+            $table->unsignedBigInteger('address_id')->after('user_id')->index();
+            $table->unsignedBigInteger('shipping_method_id')->after('address_id')->index();
 
             $table->foreign('address_id')->references('id')->on('addresses');
             $table->foreign('shipping_method_id')->references('id')->on('shipping_methods');
